comment tests

+++ controllers/commentRouter.js
@@ -54,11 +54,21 @@ commentRouter.post('/', async (req, res, next) => {
 })
 
 commentRouter.delete('/:id', async (req, res, next) => {
-  try {
-    const comment = await Comment.findByIdAndRemove(req.params.id)
-    res.status(204).end()
-  } catch (exception) {
-    next(exception)
+  if (req.get('authorization')) {
+    const token = getTokenFrom(req)
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (token && decodedToken.id) {
+      try {
+        const comment = await Comment.findByIdAndRemove(req.params.id)
+        res.status(204).end()
+      } catch (exception) {
+        next(exception)
+      }
+    } else {
+     res.status(404).end()
+    }
+  } else {
+    res.status(401).end()
   }
 })
